fix: pulled out functionality in order to use DelegationNode constructor
