+++ src/delegation/Delegation.ts
@@ -15,21 +15,13 @@
  */
 
 import { SubmittableResult } from '@polkadot/api'
-import { factory } from '../config/ConfigLog'
 import Identity from '../identity/Identity'
-import { CodecWithId, decodeDelegationNode } from './DelegationDecoder'
 import Attestation from '../attestation/Attestation'
 import { IDelegationBaseNode } from '../types/Delegation'
-import DelegationNode from './DelegationNode'
 import DelegationRootNode from './DelegationRootNode'
-import {
-  getAttestationHashes,
-  fetchChildren,
-  getChildIds,
-} from './Delegation.chain'
+import { getAttestationHashes } from './Delegation.chain'
 import { query } from '../attestation/Attestation.chain'
-
-const log = factory.getLogger('DelegationBaseNode')
+import DelegationNode from './DelegationNode'
 
 export default abstract class DelegationBaseNode
   implements IDelegationBaseNode {
@@ -70,22 +62,7 @@ export default abstract class DelegationBaseNode
    *
    * @returns Promise containing the resolved children nodes.
    */
-  public async getChildren(): Promise<DelegationNode[]> {
-    log.info(` :: getChildren('${this.id}')`)
-    const childIds: string[] = await getChildIds(this.id)
-    const queryResults: CodecWithId[] = await fetchChildren(childIds)
-    const children: DelegationNode[] = queryResults
-      .map((codec: CodecWithId) => {
-        const decoded = decodeDelegationNode(codec.codec)
-        if (decoded) return { ...decoded, id: codec.id } as DelegationNode
-        return null
-      })
-      .filter((value): value is DelegationNode => {
-        return value !== null
-      })
-    log.info(`children: ${JSON.stringify(children)}`)
-    return children
-  }
+  public abstract async getChildren(): Promise<DelegationNode[]>
 
   /**
    * Fetches and resolves all attestations attested with this delegation node.
